chore (refactoring): Move BermudaLogSpamLess to own file.
